task 7.4 - correction_2

+++ kodilla-advanced-tests/src/main/java/com/kodilla/execution_model/homework/Order.java
@@ -14,7 +14,7 @@ public class Order {
         this.login = login;
     }
 
-    public double getValue() {
+    public int getValue() {
         return value;
     }
